pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "notion-git-sync"
version = "0.1.3"
description = "Synchronize Notion pages to markdown files with frontmatter support"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["notion", "markdown", "documentation", "sync"]
authors = [
    {name = "Joe Southin", email = "joe.southin@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup"
]
dependencies = [
    "notion-client>=2.0.0",
    "pyyaml>=6.0.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0"
]
dev = [
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0"
]

[project.urls]
Homepage = "https://github.com/jsouthin/notion-git-sync"
Repository = "https://github.com/jsouthin/notion-git-sync.git"
Documentation = "https://github.com/jsouthin/notion-git-sync#readme"

[project.scripts]
notion-git-sync = "notion_git_sync.sync:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=notion_git_sync"

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
